+++ tests/emphasis-short.spec.ts
@@ -1,0 +1,223 @@
+// tslint:disable-next-line:import-name
+import dedent from 'ts-dedent';
+import { SpeechMarkdown } from '../src/SpeechMarkdown';
+
+describe('emphasis-short-moderate', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A +moderate+ level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="moderate">moderate</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="moderate">moderate</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A moderate level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});
+
+describe('emphasis-short-strong', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A ++strong++ level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="strong">strong</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="strong">strong</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A strong level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});
+
+describe('emphasis-short-none', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A ~none~ level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="none">none</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="none">none</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A none level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});
+
+describe('emphasis-short-reduced', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    A -reduced- level
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="reduced">reduced</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      A <emphasis level="reduced">reduced</emphasis> level
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      A reduced level
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});