+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to the speechmarkdown-js project will be documented in this file.
 
+## 0.3.0-alpha.0 - (June 30, 2019)
+### Added
+- Update grammar and formatters for emphasis short format
+- Change speechmarkdown.toString(markdown) to speechmarkdown.toText(markdown)
+
+
 ## 0.2.0-alpha.0 - (June 29, 2019)
 ### Added
 - CHANGELOG.md